+++ RoomArrangement/CardinalDirections.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace RoomArrangement
+{
+	[Flags]
+	public enum CardinalDirections
+	{
+		None = 0,
+		North = 1 << 0,
+		East = 1 << 1,
+		South = 1 << 2,
+		West = 1 << 3
+	}
+}
